weather: add unit parameter to commands
The unit parameter specifies if the weather data should be displayed
in imperial or metric units. It is an optional argument which defaults
to metric units. The argument is passed using the flag -u/--unit.

The valid values are 'c' and 'f', the first letters of the temperature
units associated with metric and imperial, respectively.